Implement first cut DCR in SD
git-svn-id: https://bacula.svn.sourceforge.net/svnroot/bacula/trunk@1113 91ce42f0-d328-0410-95d8-f526ca767f89

+++ bacula/src/lib/message.c
@@ -380,10 +380,10 @@ static void make_unique_mail_filename(JCR *jcr, POOLMEM **name, DEST *d)
 {
    if (jcr) {
       Mmsg(name, "%s/%s.mail.%s.%d", working_directory, my_name,
-		 jcr->Job, (int)d);
+		 jcr->Job, (int)(long)d);
    } else {
       Mmsg(name, "%s/%s.mail.%s.%d", working_directory, my_name,
-		 my_name, (int)d);
+		 my_name, (int)(long)d);
    }
    Dmsg1(200, "mailname=%s\n", *name);
 }
